proyecto_final/requirements.txt
# Dependencias para el análisis de patrones de interferencia
# Proyecto: Determinación de la Velocidad de la Luz - Interferómetro de Michelson

# Procesamiento numérico
numpy>=1.21.0

# Procesamiento de señales y análisis científico
scipy>=1.7.0

# Visualización y gráficos
matplotlib>=3.4.0

# Procesamiento de imágenes
Pillow>=8.3.0
opencv-python>=4.5.0

# Utilidades opcionales
# pandas>=1.3.0          # Para análisis de datos tabulares
# jupyter>=1.0.0         # Para notebooks interactivos
# scikit-image>=0.18.0   # Procesamiento avanzado de imágenes
